test: isolate consensus algorithm tests
